Improve TypeScript types and add strict type checking
Changes:
- Created centralized types.ts files at appropriate levels:
  - src/types.ts: Shared interface contracts between core and adapters
  - src/core/types.ts: Core domain types (game logic, difficulty, scoring)
  - src/adapters/types.ts: Adapter-specific types (input, sound, terminal)

- Updated tsconfig.json with stricter type checking:
  - Added noUncheckedIndexedAccess for safer array/object access
  - Added noImplicitReturns to ensure all code paths return
  - Added noFallthroughCasesInSwitch for switch statement safety
  - Removed bun-types from types array (causes tsc issues)

- Fixed type safety issues found by stricter settings:
  - Added undefined checks for array/object access
  - Fixed potential undefined values in game-logic.ts
  - Fixed array access issues in word-pool.ts and word-generation.ts
  - Fixed level access in level-progression.ts
  - Fixed NodeJS.Timeout type usage in game.ts

- Changed all import extensions from .js to .ts throughout codebase

- Added check script to package.json:
  - Runs both type-check and tests
  - Provides single command for full validation

- Created GitHub Actions workflow (.github/workflows/type-check.yml):
  - Runs on pushes to main/master and on pull requests
  - Installs dependencies with Bun
  - Runs type checking
  - Runs tests

All types are now properly organized and the codebase has significantly
improved type safety with stricter TypeScript compiler options.
