+++ src/adapters/real-sound.ts
@@ -3,8 +3,8 @@
  * Wraps the existing sound-adapter for use with the game
  */
 
-import type { SoundPlayer } from '../core/game.js';
-import { createSoundAdapter } from './sound-adapter.js';
+import type { SoundPlayer } from '../types.ts';
+import { createSoundAdapter } from './sound-adapter.ts';
 
 export function createRealSound(): SoundPlayer {
 	const sound = createSoundAdapter();
